backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests
pydantic
openai
python-dotenv
volcengine-python-sdk[ark]
redis
